Fix various issues with keyboard controllers and state machine.
Combined handlers for wait_for_first_player and wait_for_next_player
since most of the code is the same, and to avoid having these diverge.

NEXT: work on antenna sketch to handle commands, similar to the keyboard
controller and make the antenna controller a listener.

NEXT: add new listener methods to GUI and get that working.