More initialization adjustments
Modifications to the controller initialization to more reliably detect
connection to a Keyboard or Antenna.  Tested case where the wrong type
was tried first.

Added quit command to gracefully exit.  Also trap interrupts and
gracefully exit (to avoid RX TX Warnings about stale lock files on
subsequent startup).

Stop program if we are not able to connect to all of the requested
controllers.