Optimize initialization with multiple antennas.
If the system rejects a port when looking for an antenna,
mark it, so we do not try it when opening subsequent antennas.

Modified Antenna Sketch to keep announcing that it could not find the
RFID chip, rather than quietly locking up, so we do not hang waiting for
and ident message.