Keybaord controller updates:
1) Add "no cards played" to start of status announcement, when
appropriate
2) Change getHand to return my hand if asking for dummy and I am dummy
(not used because of change #3)
3) Announce "Your partner is playing your hand" when clicking dummy suit
buttons and you are dummy.
4) Automatically select singleton card only when selecting a suit for
the player to play next (me or dummy, if I am playing dummy).

Fix to Game Controller:
1) clear dummy hand when auto scanning dummy from test hand and dummy
has a keyboard controller (so that we don't announce hand complete with
each card "scanned").  We still need to scan the dummy in case there are
other keyboard controllers.

2) add support for "A" as a direction for the command line "printhand"
command, which prints all known hands.  So, you can now type "ph a" or
"pha" to see all hands during testing.